[#126] Rename the `apigee_subscribe_form` and `apigee_subscribe_link` formatters.
