+++ config/install/core.entity_view_display.rate_plan.rate_plan.default.yml
@@ -24,7 +24,7 @@ content:
     settings: {  }
     third_party_settings: {  }
   subscribe:
-    type: apigee_subscribe_link
+    type: apigee_purchase_plan_link
     weight: 2
     region: content
     label: visually_hidden
