+++ config/schema/apigee_m10n.schema.yml
@@ -46,21 +46,21 @@ apigee_m10n.purchased_plan.config:
       type: label
       label: 'Already purchased title.'
 
-field.formatter.settings.apigee_subscribe_link:
+field.formatter.settings.apigee_purchase_plan_link:
   type: mapping
-  label: 'Subscribe Link'
+  label: 'Purchase plan link'
   mapping:
     label:
       type: label
-      label: 'Subscribe label'
+      label: 'Purchase label'
 
-field.formatter.settings.apigee_subscribe_form:
+field.formatter.settings.apigee_purchase_plan_form:
   type: mapping
-  label: 'Subscribe form settings'
+  label: 'Rate plan purchase form settings'
   mapping:
     label:
       type: label
-      label: 'Subscribe Label'
+      label: 'Purchase Label'
 
 field.formatter.settings.apigee_datestamp:
   type: mapping
