+++ src/Plugin/Field/FieldFormatter/PurchasePlanFormFormatter.php
@@ -19,29 +19,38 @@
 
 namespace Drupal\apigee_m10n\Plugin\Field\FieldFormatter;
 
+use Apigee\Edge\Api\Monetization\Entity\Developer;
+use Drupal\apigee_m10n\Entity\PurchasedPlan;
+use Drupal\Core\Entity\EntityFormBuilderInterface;
+use Drupal\Core\Field\FormatterBase;
+use Drupal\Core\Field\FieldItemListInterface;
+use Drupal\Core\Field\FieldItemInterface;
 use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Drupal\Core\Field\FieldDefinitionInterface;
-use Drupal\Core\Field\FieldItemListInterface;
-use Drupal\Core\Field\FormatterBase;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\Core\Field\FieldItemInterface;
-use Drupal\Core\Link;
 use Drupal\apigee_m10n\Monetization;
 use Drupal\apigee_m10n\Form\PurchasedPlanConfigForm;
 
 /**
- * Plugin implementation of the 'apigee_subscription_form' formatter.
+ * Plugin implementation of the 'apigee_purchase_plan_form' formatter.
  *
  * @FieldFormatter(
- *   id = "apigee_subscribe_link",
- *   label = @Translation("Link to form"),
+ *   id = "apigee_purchase_plan_form",
+ *   label = @Translation("Rendered form"),
  *   field_types = {
  *     "apigee_subscribe"
  *   }
  * )
  */
-class SubscribeLinkFormatter extends FormatterBase implements ContainerFactoryPluginInterface {
+class PurchasePlanFormFormatter extends FormatterBase implements ContainerFactoryPluginInterface {
+
+  /**
+   * Entity form builder.
+   *
+   * @var \Drupal\Core\Entity\EntityFormBuilderInterface
+   */
+  protected $entityFormBuilder;
 
   /**
    * The Cache backend.
@@ -51,7 +60,7 @@ class SubscribeLinkFormatter extends FormatterBase implements ContainerFactoryPl
   private $monetization;
 
   /**
-   * Creates an instance of the plugin.
+   * SubscribeLinkFormatter constructor.
    *
    * @param string $plugin_id
    *   The plugin_id for the formatter.
@@ -66,12 +75,15 @@ class SubscribeLinkFormatter extends FormatterBase implements ContainerFactoryPl
    * @param string $view_mode
    *   The view mode.
    * @param array $third_party_settings
+   *   Any third party settings.
+   * @param \Drupal\Core\Entity\EntityFormBuilderInterface $entityFormBuilder
    *   Entity form builder service.
    * @param \Drupal\apigee_m10n\Monetization $monetization
    *   Monetization service.
    */
-  public function __construct($plugin_id, $plugin_definition, FieldDefinitionInterface $field_definition, array $settings, $label, $view_mode, array $third_party_settings, Monetization $monetization) {
+  public function __construct($plugin_id, $plugin_definition, FieldDefinitionInterface $field_definition, array $settings, $label, $view_mode, array $third_party_settings, EntityFormBuilderInterface $entityFormBuilder, Monetization $monetization) {
     parent::__construct($plugin_id, $plugin_definition, $field_definition, $settings, $label, $view_mode, $third_party_settings);
+    $this->entityFormBuilder = $entityFormBuilder;
     $this->monetization = $monetization;
   }
 
@@ -87,6 +99,7 @@ class SubscribeLinkFormatter extends FormatterBase implements ContainerFactoryPl
       $configuration['label'],
       $configuration['view_mode'],
       $configuration['third_party_settings'],
+      $container->get('entity.form_builder'),
       $container->get('apigee_m10n.monetization')
     );
   }
@@ -106,7 +119,7 @@ class SubscribeLinkFormatter extends FormatterBase implements ContainerFactoryPl
   public function settingsForm(array $form, FormStateInterface $form_state) {
     $form['label'] = [
       '#type' => 'textfield',
-      '#title' => $this->t('Subscribe label'),
+      '#title' => $this->t('Button label'),
       '#default_value' => $this->getSetting('label'),
     ];
     return $form;
@@ -124,19 +137,22 @@ class SubscribeLinkFormatter extends FormatterBase implements ContainerFactoryPl
   }
 
   /**
-   * Renderable link element.
+   * Renderable entity form.
    *
    * @param \Drupal\Core\Field\FieldItemInterface $item
    *   Field item variable.
    *
    * @return array
-   *   Renderable link element.
+   *   Renderable form elements.
+   *
+   * @throws \Exception
    */
   protected function viewValue(FieldItemInterface $item) {
     /** @var \Drupal\apigee_m10n\Entity\RatePlanInterface $rate_plan */
     $rate_plan = $item->getEntity();
-    if ($value = $item->getValue()) {
-      if ($this->monetization->isDeveloperAlreadySubscribed($value['user']->getEmail(), $rate_plan)) {
+    if (($value = $item->getValue()) && $item->getEntity()->access('subscribe')) {
+      $developer_id = $value['user']->getEmail();
+      if ($this->monetization->isDeveloperAlreadySubscribed($developer_id, $rate_plan)) {
         $label = \Drupal::config(PurchasedPlanConfigForm::CONFIG_NAME)->get('already_purchased_label');
         return [
           '#markup' => $this->t($label ?? 'Already purchased %rate_plan', [
@@ -144,14 +160,19 @@ class SubscribeLinkFormatter extends FormatterBase implements ContainerFactoryPl
           ])
         ];
       }
-
-      return Link::createFromRoute(
-        $this->getSetting('label'), 'entity.rate_plan.subscribe', [
-          'user'      => $value['user']->id(),
-          'package'   => $rate_plan->getPackage()->id(),
-          'rate_plan' => $rate_plan->id(),
-        ]
-      )->toRenderable();
+      $start_date = new \DateTimeImmutable();
+      $org_timezone = $rate_plan->getOrganization()->getTimezone();
+      $start_date->setTimezone($org_timezone);
+      $purchased_plan = PurchasedPlan::create([
+        'ratePlan' => $rate_plan,
+        // TODO: User a controller proxy that caches the developer entity.
+        // @see: https://github.com/apigee/apigee-edge-drupal/pull/97.
+        'developer' => new Developer(['email' => $developer_id]),
+        'startDate' => $start_date,
+      ]);
+      return $this->entityFormBuilder->getForm($purchased_plan, 'default', [
+        'save_label' => $this->t('@save_label', ['@save_label' => $this->getSetting('label')]),
+      ]);
     }
   }
